pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "robothor"
version = "0.1.0"
description = "An AI brain with persistent memory, semantic search, knowledge graph, vision, and self-healing infrastructure."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Robothor AI", email = "robothor@ironsail.ai" },
]
keywords = ["ai", "brain", "memory", "knowledge-graph", "rag", "vision", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "psycopg2-binary>=2.9",
    "redis>=5.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
]
vision = [
    "ultralytics>=8.0",
    "insightface>=0.7",
    "onnxruntime>=1.17",
]
llm = [
    "ollama>=0.4",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.8",
    "mypy>=1.13",
]
all = [
    "robothor[api,vision,llm,dev]",
]

[project.scripts]
robothor = "robothor.cli:main"

[project.urls]
Homepage = "https://robothor.ai"
Repository = "https://github.com/robothor-ai/robothor"
Issues = "https://github.com/robothor-ai/robothor/issues"

[tool.hatch.build.targets.wheel]
packages = ["robothor"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires real database/Redis",
    "llm: requires Ollama running",
    "slow: takes more than 10 seconds",
    "e2e: full system end-to-end",
    "smoke: health check only",
]
asyncio_mode = "auto"
